Combine FormInput components
